requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
docker>=6.1.0
kafka-python>=2.0.2
httpx>=0.25.0
python-multipart>=0.0.6

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
psutil>=5.9.0
pyyaml>=6.0